requirements.txt
# Core LangChain dependencies
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0

# Vector stores and embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.0
pinecone-client>=3.0.0

# Document processing
pypdf>=4.0.0
docx2txt>=0.8
unstructured>=0.10.0

# Web scraping and APIs
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Async support
aiohttp>=3.8.0
aiofiles>=23.0.0

# Development and testing
jupyter>=1.0.0
ipykernel>=6.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0
pydantic>=2.0.0

# Optional: Additional model providers
# openai>=1.0.0
# anthropic>=0.25.0
# google-generativeai>=0.5.0
